rename SimpleRenderer to just Renderer and move it to mod
